Bug 534459 - Readme for 4.8
Ackknowledge for the Eclipse Platform + to Holger for this Eclipse
promotion videos.

As I'm not aware of an automatic way to retrieve the names of the active
committers, he also thank them which I think is also a nice gesture, not
all of us are getting paid to work on Eclipse.

Change-Id: Ieca26bfacc3c8286b9adf8975673cb79a182857e
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>