Bug 514294 - an Eclipse view developed with HTML/JS and Browser widget
In order to easily demo the Browser widget and how it can allow to use
HTML/JS to create Eclipse components such as views, templates were
developed to create integrated views relying on the Browser Technology
and showing operations such as being synchronized with the
SelectionService or something that shows the view can interact in both
ways with the IDE.

In addition, a template was created to show how to run a standalone web
application in the form of a brick game.

The views can be created using a workbench with PDE only, but support
for web technology like the Wild Web Developer plug-in is recommended.


Change-Id: If4a4a18a37c5e133ce14b7e9e7e16213625a4bad